[players] Move annotation libs under src/lib/players
annotation.js, arrowshape.js and eraserbrush.js are all player-only
annotation helpers; group them under src/lib/players/ next to the
composables that consume them. Update the import paths in the two
annotation components, the two annotation composables and the three
specs. The cross-imports between the three files are relative, so they
keep working unchanged.

Also rewrite annotation.js's header: it no longer narrates the old
Options-API mixin extraction, just describes what the module provides
today (shared stateless annotation helpers) and its import-time side
effects (Arrow registration + eraser prototype patch).
