+++ src/lib/players/eraserbrush.js
